% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5_dataset.R
\name{h5Dataset}
\alias{h5Dataset}
\title{Read hdf5 file data}
\usage{
h5Dataset(file, name, bit64conversion = TRUE, transpose = FALSE, raw = FALSE)
}
\description{
Read hdf5 file data, this function returns ALTREP objects
to represent the dataset in R.
}
