\name{HighFive-package}
\alias{HighFive-package}
\alias{HighFive}
\docType{package}
\title{
  The HighFive C++ interface to HDF5
}
\description{
  HighFive is a modern, user-friendly, header-only, C++14 interface for libhdf5.
}
\details{
  HighFive is a modern, user-friendly, header-only, C++14 interface for libhdf5.

  Bug reports can also be registered at the GitHub issue tracker at
  \url{https://github.com/theAeon/HighFive/issues}.
}
\author{
  Andrew Robbins

  Maintainer: Andrew Robbins <robbiand@umich.edu>
}
\references{
  \url{https://github.com/highfive-devs/highfive/}
}
\keyword{ package }
\examples{
  # None
}
